A days work
Code to filter out mutants which might not be deleterious according to VEP metrics

DEA run 6 times before and after this change to see if it makes any difference - looks like it probably doesn't? Could do with running more times and storing the number of sig genes in a useful location.

Improved code headings